fix: Mixing flattened maps
