+++ test/unflatten_test.dart
@@ -2,32 +2,69 @@ import 'package:flat/flat.dart';
 import 'package:test/test.dart';
 
 void main() {
-  test('Unflattens map', () {
-    const obj = {
-      "a": 1,
-      "list1.0": "item1",
-      "list1.1": "item2",
-      "f.list2.0": "item3",
-      "f.list2.1": "item4",
-      "f.list2.2": "item5",
-      "f.g": 2,
-      "f.h": true,
-      "f.j": "text"
-    };
+  group('unflatten', () {
+    test('Nested once', () {
+      const obj = {
+        'hello.world': 'good morning',
+      };
 
-    const expected = {
-      "a": 1,
-      "list1": ["item1", "item2"],
-      "f": {
-        "list2": ["item3", "item4", "item5"],
-        "g": 2,
-        "h": true,
-        "j": "text",
-      },
-    };
+      const expected = {
+        "hello": {
+          "world": "good morning",
+        },
+      };
 
-    final result = unflatten(obj);
+      final result = unflatten(obj);
 
-    expect(result, expected);
+      expect(result, expected);
+    });
+
+    test('Nested twice', () {
+      const obj = {
+        'hello.world.again': 'good morning',
+      };
+
+      const expected = {
+        "hello": {
+          "world": {
+            "again": 'good morning',
+          },
+        }
+      };
+
+      final result = unflatten(obj);
+
+      expect(result, expected);
+    });
+
+    test('Multiple Keys', () {
+      const obj = {
+        'hello.lorem.ipsum': 'again',
+        'hello.lorem.dolor': 'sit',
+        'world.lorem.ipsum': 'again',
+        'world.lorem.dolor': 'sit',
+        'world': {'greet': 'hello'},
+      };
+
+      const expected = {
+        "hello": {
+          "lorem": {
+            'ipsum': 'again',
+            'dolor': 'sit',
+          },
+        },
+        'world': {
+          'greet': 'hello',
+          'lorem': {
+            'ipsum': 'again',
+            'dolor': 'sit',
+          },
+        },
+      };
+
+      final result = unflatten(obj);
+
+      expect(result, expected);
+    });
   });
 }
